% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{add_data}
\alias{add_data}
\title{Add data to a sgldata data object}
\usage{
add_data(data, x, name, ...)
}
\arguments{
\item{data}{sgldata object}

\item{x}{data object to add,}

\item{name}{name of data object}

\item{...}{additional parameters}
}
\description{
Addes a data to a sgldata data object
}
\author{
Martin Vincent
}
