% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{add_data.sgldata}
\alias{add_data.sgldata}
\title{Add data to a sgldata data object}
\usage{
\method{add_data}{sgldata}(data, x, name, default = NULL,
  type = element_class(x), sparse = is(x, "sparseMatrix"), ...)
}
\arguments{
\item{data}{sgldata object}

\item{x}{matrix or vector,}

\item{name}{name of the data object}

\item{default}{default value, returned if x is null}

\item{type}{data type, 'numeric' or 'integer'}

\item{sparse}{if TRUE \code{y} will be treated as sparse, if FALSE \code{y} will be treated as dens.}

\item{...}{not used}
}
\description{
Addes a data to a sgldata data object
}
\seealso{
Other sgldata: \code{\link{create.sgldata}},
  \code{\link{prepare.args.sgldata}},
  \code{\link{prepare.args}}, \code{\link{prepare_data}},
  \code{\link{rearrange.sgldata}},
  \code{\link{subsample.sgldata}}
}
\author{
Martin Vincent
}
\concept{sgldata}
