% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.kron}
\alias{as.kron}
\title{Convert a list of matrices to a kron object}
\usage{
as.kron(x)
}
\arguments{
\item{x}{a list of matrices}
}
\value{
a kron object
}
\description{
Convert a list of matrices to a kron object
}
\author{
martin
}

