% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{best_model}
\alias{best_model}
\title{Index of best model}
\usage{
best_model(object, ...)
}
\arguments{
\item{object}{an sgl object}

\item{...}{additional parameters (optional)}
}
\value{
index of the best model.
}
\description{
Returns the index of the best model
}
\author{
Martin Vincent
}

