% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{best_model.sgl}
\alias{best_model.sgl}
\title{Index of best model}
\usage{
\method{best_model}{sgl}(object, pkg, ...)
}
\arguments{
\item{object}{a sgl object}

\item{pkg}{name of calling package}

\item{...}{additional parameters (optional)}
}
\value{
index of the best model.
}
\description{
Returns the index of the best model, in terms of lowest error rate
}
\author{
Martin Vincent
}

