% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{coef.sgl}
\alias{coef.sgl}
\title{Extracting the nonzero coefficients}
\usage{
\method{coef}{sgl}(object, index = 1:nmod(object), ...)
}
\arguments{
\item{object}{a sgl object}

\item{index}{indices of the models}

\item{...}{not used}
}
\value{
a list of with nonzero coefficients of the models
}
\description{
Extracting the nonzero coefficients
}
\author{
Martin Vincent
}

