% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_error.R
\name{compute_error}
\alias{compute_error}
\title{Helper function for computing error rates}
\usage{
compute_error(x, response_name, true_response, loss,
  transposed_response = FALSE)
}
\arguments{
\item{x}{sgl object containing responses}

\item{response_name}{the name of the response.}

\item{true_response}{the true response}

\item{loss}{the loss function.}

\item{transposed_response}{have the response list been transposed with \code{transpose_response_elements}}
}
\value{
a vector or matrrix with the computed error rates
}
\description{
This function can be used to compute mean error rates.
It is consist with the use cases of the \code{Err} genetic function.
The loss function should be of the form
\code{function(x, y)}
and must return a single numeric number, with x a list of true responses and y a list of responses (one list element for each sample).
}
\author{
Martin Vincent
}
