% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{element_class}
\alias{element_class}
\title{Retur the element class of an object.}
\usage{
element_class(x)
}
\arguments{
\item{x}{a matrix, vector or NULL}
}
\description{
Return the element class of an object.
The object must be a matrix, vector or NULL.
The element class of NULL is NULL
}
\author{
Martin Vincent
}

