% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_error.R
\name{Err.sgl}
\alias{Err.sgl}
\title{Error Rates}
\usage{
\method{Err}{sgl}(object, data = NULL, response = object$Y.true, ...)
}
\arguments{
\item{object}{a lsgl object.}

\item{data}{a design matrix (the \eqn{X} matrix).}

\item{response}{a matrix of the true responses (the \eqn{Y} matrix).}

\item{...}{ignored.}
}
\value{
a vector of errors.
}
\description{
Compute and return the root-mean-square error for each model.
This method is only intended for testing.

The root-mean-square error (RMSE) is
\deqn{\frac{1}{K}\sum_{i = 1}^K \sqrt{\frac{1}{N}\sum_{j=1}^N Y_{ji}-(X\hat \beta)_{ji}}}
RMSE is the default error.
}
\author{
Martin Vincent
}

