% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{features}
\alias{features}
\title{Extracts nonzero features}
\usage{
features(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters (optional)}
}
\value{
a list of length \code{nmod(x)} containing the nonzero features of the models.
}
\description{
Generic function for extracting nonzero features.
}
\author{
Martin Vincent
}
