% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{features_stat}
\alias{features_stat}
\title{Extract feature statistics}
\usage{
features_stat(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters (optional)}
}
\value{
an object conting the computed statistics.
}
\description{
Generic function for extracting feature statistics.
}
\author{
Martin Vincent
}
