% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{get_coef}
\alias{get_coef}
\title{Get the nonzero coefficients}
\usage{
get_coef(object, index = 1:length(object))
}
\arguments{
\item{object}{a list of lists of matrices or a list of matrices}

\item{index}{indices to be extracted from}
}
\value{
a list (of lists) with the nonzero coefficients
}
\description{
Extracting nonzero coefficients from list (of lists) of matrices
}
\author{
Martin Vincent
}
