% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kron}
\alias{kron}
\title{Create Kronecker product representation}
\usage{
kron(...)
}
\arguments{
\item{...}{matrices in the product}
}
\value{
Kronecker product representation
}
\description{
Create Kronecker product representation
}
\author{
Martin Vincent
}

