% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_for_testing.R
\name{linear_test_diagonal_error_w_sgl_lambda_R}
\alias{linear_test_diagonal_error_w_sgl_lambda_R}
\title{C interface}
\usage{
linear_test_diagonal_error_w_sgl_lambda_R(data, block_dim, groupWeights,
  parameterWeights, alpha, d, lambda.min, lambda.min.rel, algorithm.config)
}
\description{
C interface
}
\keyword{internal}
