% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_for_testing.R
\name{linear_test_diagonal_w_spx_spy_sgl_test_R}
\alias{linear_test_diagonal_w_spx_spy_sgl_test_R}
\title{C interface}
\usage{
linear_test_diagonal_w_spx_spy_sgl_test_R(data, block_dim, groupWeights,
  parameterWeights, algorithm.config)
}
\description{
C interface
}
\keyword{internal}
