% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_for_testing.R
\name{linear_test_identity_spx_sgl_subsampling_R}
\alias{linear_test_identity_spx_sgl_subsampling_R}
\title{C interface}
\usage{
linear_test_identity_spx_sgl_subsampling_R(data, test_data, block_dim,
  groupWeights, parameterWeights, alpha, lambda, idx, algorithm.config)
}
\description{
C interface
}
\keyword{internal}
