% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{models}
\alias{models}
\title{Extract fitted models}
\usage{
models(object, index, ...)
}
\arguments{
\item{object}{an object}

\item{index}{a vector of indices of the models to be returned}

\item{...}{additional parameters (optional)}
}
\value{
a list of length \code{length(index)} containing the models
}
\description{
Generic function for extracting the fitted models.
Returns the fitted models.
}
\author{
Martin Vincent
}
