% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{models.sgl}
\alias{models.sgl}
\title{Extract the estimated models}
\usage{
\method{models}{sgl}(object, index = 1:nmod(object), ...)
}
\arguments{
\item{object}{a sgl object}

\item{index}{indices of the models to be returned}

\item{...}{ignored}
}
\value{
a list of sparse matrices
}
\description{
This function returns the estimated models (that is the \eqn{beta} matrices)
}
\author{
Martin Vincent
}

