% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{nmod.sgl}
\alias{nmod.sgl}
\title{Returns the number of models in a sgl object}
\usage{
\method{nmod}{sgl}(object, ...)
}
\arguments{
\item{object}{a sgl object}

\item{...}{ignored}
}
\value{
the number of models in \code{object}
}
\description{
Returns the number of models used for fitting.
}
\details{
Note that cv and subsampling objects does not conating any models even though nmod returns a positiv number.
}
\author{
Martin Vincent
}
