% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{parameters}
\alias{parameters}
\title{Extracts nonzero parameters}
\usage{
parameters(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters (optional)}
}
\value{
a list of length \code{nmod(x)} containing the nonzero parameters of the models.
}
\description{
Generic function for extracting nonzero parameters for each model.
}
\author{
Martin Vincent
}
