% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{parameters_stat}
\alias{parameters_stat}
\title{Extract parameter statistics}
\usage{
parameters_stat(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters (optional)}
}
\value{
an object conting the computed statistics.
}
\description{
Generic function for extracting parameter statistics.
}
\author{
Martin Vincent
}
