% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{parameters_stat.sgl}
\alias{parameters_stat.sgl}
\title{Extracting parameter statistics}
\usage{
\method{parameters_stat}{sgl}(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{ignored}
}
\value{
a vector of length \code{nmod(x)} or a matrix containing the number of nonzero parameters of the models.
}
\description{
Extracts the number of nonzero parameters in each model.
}
\author{
Martin Vincent
}

