% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare a sgldata data object}
\usage{
prepare_data(x, default = NULL, type = "numeric", sparse = is(x,
  "sparseMatrix"))
}
\arguments{
\item{x}{the matrix,}

\item{default}{default value, returned if x is null}

\item{type}{data type, 'numeric' or 'integer'}

\item{sparse}{if TRUE \code{y} will be treated as sparse, if FALSE \code{y} will be treated as dens.}
}
\description{
Creates a sgldata data object from a matrix or vector
}
\author{
Martin Vincent
}
\seealso{
Other sgldata: \code{\link{add_data.sgldata}},
  \code{\link{create.sgldata}},
  \code{\link{prepare.args.sgldata}},
  \code{\link{prepare.args}},
  \code{\link{rearrange.sgldata}},
  \code{\link{subsample.sgldata}}
}

