% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{print_with_metric_prefix}
\alias{print_with_metric_prefix}
\title{Print a numeric with metric prefix}
\usage{
print_with_metric_prefix(x, digits = 3)
}
\arguments{
\item{x}{numeric to be printed}

\item{digits}{number of significant digits}
}
\value{
a string
}
\description{
Print a numeric with metric prefix
}
\author{
Martin Vincent
}
