% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{rearrange}
\alias{rearrange}
\title{Generic rearrange function}
\usage{
rearrange(data, covariate.order, ...)
}
\arguments{
\item{data}{a data object}

\item{covariate.order}{the new order of the covariates}

\item{...}{additional parameters}
}
\value{
a rearranged data object of same class as \code{data}
}
\description{
Rearrange the order of the covariates in the \code{data} object.
}
\seealso{
rearrange.sgldata
}
\author{
Martin Vincent
}
