% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_config.R
\name{sgl.c.config}
\alias{sgl.c.config}
\title{Featch information about the C side configuration of the package}
\usage{
sgl.c.config()
}
\value{
list
}
\description{
Featch information about the C side configuration of the package
}
\author{
Martin Vicnet
}

