% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_config.R
\docType{data}
\name{sgl.standard.config}
\alias{sgl.standard.config}
\title{Standard algorithm configuration}
\format{An object of class \code{list} of length 15.}
\usage{
sgl.standard.config
}
\description{
\code{sgl.standard.config <- sgl.algorithm.config()}
}
\author{
Martin Vicnet
}
\keyword{datasets}

