% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{sgl_print}
\alias{sgl_print}
\title{Print information about sgl object}
\usage{
sgl_print(x)
}
\arguments{
\item{x}{a object of sgl family class}
}
\description{
Prints information about sgl object
}
\author{
Martin Vincent
}

