% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{sparseMatrix_from_C_format}
\alias{sparseMatrix_from_C_format}
\title{Convert to sparse matrix}
\usage{
sparseMatrix_from_C_format(x)
}
\arguments{
\item{x}{.Call retunred list}
}
\description{
Convert sparse matrix  retunred from .Call to sparseMatrix.
}
\author{
Martin Vincent
}

