% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{sparseMatrix_to_C_format}
\alias{sparseMatrix_to_C_format}
\title{Prepare sparse matrix for .Call}
\usage{
sparseMatrix_to_C_format(x)
}
\arguments{
\item{x}{a spares matrix}
}
\description{
Prepare sparse matrix for .Call
}
\author{
Martin Vincent
}
