% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{subsample}
\alias{subsample}
\title{Subsample}
\usage{
subsample(data, indices, ...)
}
\arguments{
\item{data}{a data object}

\item{indices}{a vector of indices to pick out}

\item{...}{not used}
}
\value{
a data object of the same class as \code{data}
}
\description{
Pick out a subsample of an object
}
\author{
Martin Vincent
}
