% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_args.R
\name{subsample.sgldata}
\alias{subsample.sgldata}
\title{Subsample sgldata}
\usage{
\method{subsample}{sgldata}(data, indices, ...)
}
\arguments{
\item{data}{a sgldata object}

\item{indices}{a vector of indices to pick out}

\item{...}{not used}
}
\value{
a sgldata
}
\description{
Pick out a subsample of a sgldata object
}
\seealso{
Other sgldata: \code{\link{add_data.sgldata}},
  \code{\link{create.sgldata}},
  \code{\link{prepare.args.sgldata}},
  \code{\link{prepare.args}}, \code{\link{prepare_data}},
  \code{\link{rearrange.sgldata}}
}
\author{
Martin Vincent
}
