% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools_test.R
\name{test_rtools}
\alias{test_rtools}
\title{Test internal rtools}
\usage{
test_rtools()
}
\description{
This function runs some internal tests and is not intended for users of the package.
}
\author{
Martin Vincent
}

