% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_formatter.R
\name{transpose_response_elements}
\alias{transpose_response_elements}
\title{Transpose response elements}
\usage{
transpose_response_elements(x)
}
\arguments{
\item{x}{response list or matrix}
}
\value{
response list with all matrices transposed
}
\description{
Transpose response elements in a response list and sub lists
}
\author{
Martin Vincent
}
