% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_navigate.R
\name{coef.sgl}
\alias{coef.sgl}
\title{Extracting the nonzero coefficients}
\usage{
\method{coef}{sgl}(object, index = 1:nmod(object), parameter = "beta", ...)
}
\arguments{
\item{object}{a sgl object}

\item{index}{indices of the models}

\item{parameter}{name of the parameter (deafult is 'beta')}

\item{...}{ignored}
}
\value{
a list of with nonzero coefficients of the models
}
\description{
This function returns the nonzero coefficients
(that is the nonzero entries of the \eqn{beta} matrices)
}
\author{
Martin Vincent
}
