\name{features}
\alias{features}
\title{Generic function for extracting nonzero features (or groups)}
\usage{
  features(x, ...)
}
\arguments{
  \item{x}{an object}

  \item{...}{additional paramters (optional)}
}
\value{
  a list of nonzero features
}
\description{
  Generic function for extracting nonzero features (or
  groups)
}
\author{
  Martin Vincent
}

