\name{models}
\alias{models}
\title{Generic function for extracting the fitted models}
\usage{
  models(x, index, ...)
}
\arguments{
  \item{x}{an object}

  \item{index}{a vector of indices of the models to be
  returned}

  \item{...}{additional paramters (optional)}
}
\value{
  a list of length \code{length(index)} containing the
  models
}
\description{
  Generic function for extracting the fitted models
}
\author{
  Martin Vincent
}

