\name{nmod}
\alias{nmod}
\title{Generic function for couting the number of models}
\usage{
  nmod(x, ...)
}
\arguments{
  \item{x}{an object}

  \item{...}{additional paramters (optional)}
}
\value{
  the number of models contained in \code{x}
}
\description{
  Generic function for couting the number of models
}
\author{
  Martin Vincent
}

