% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nmod}
\alias{nmod}
\title{Generic function for counting the number of models}
\usage{
nmod(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters (optional)}
}
\value{
the number of models contained in the object \code{x}.
}
\description{
Returns the number of models
}
\author{
Martin Vincent
}

