% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nmod.sgl}
\alias{nmod.sgl}
\title{Returns the number of models in a sgl object}
\usage{
\method{nmod}{sgl}(object, ...)
}
\arguments{
\item{object}{a sgl object}

\item{...}{not used}
}
\value{
the number of models in \code{object}
}
\description{
Returns the number of models in a sgl object
}
\author{
Martin Vincent
}

