\name{parameters}
\alias{parameters}
\title{Generic function for extracting nonzero parameters}
\usage{
  parameters(x, ...)
}
\arguments{
  \item{x}{an object}

  \item{...}{additional paramters (optional)}
}
\value{
  a list of nonzero paramters
}
\description{
  Generic function for extracting nonzero parameters
}
\author{
  Martin Vincent
}

