\name{parameters.sgl}
\alias{parameters.sgl}
\title{Extracting nonzero parameters}
\usage{
  \method{parameters}{sgl} (object, ...)
}
\arguments{
  \item{object}{a sgl object}

  \item{...}{not used}
}
\value{
  a list of vectors containing the nonzero parameters (that
  is nonzero entries of the \eqn{beta} matrices)
}
\description{
  Extracting nonzero parameters
}
\author{
  Martin Vincent
}

