\name{prepare.args}
\alias{prepare.args}
\title{todo}
\usage{
  prepare.args(data, ...)
}
\arguments{
  \item{data}{a data object}

  \item{...}{additional paramters}
}
\value{
  \item{block.dim}{a vector} \item{groupWeights}{}
  \item{parameterWeights}{} \item{alpha}{} \item{data}{}
  \item{group.order}{}
}
\description{
  todo
}
\author{
  Martin Vincent
}

