\name{prepare.args}
\alias{prepare.args}
\title{Prepare sgl function arguments}
\usage{
  prepare.args(data, covariateGrouping, groupWeights,
    parameterWeights, alpha)
}
\arguments{
  \item{data}{}

  \item{covariateGrouping}{}

  \item{groupWeights}{}

  \item{parameterWeights}{}

  \item{alpha}{}
}
\description{
  Prepare sgl function arguments
}
\author{
  Martin Vincent
}

