\name{rearrange}
\alias{rearrange}
\title{todo}
\usage{
  rearrange(data, covariate.order, ...)
}
\arguments{
  \item{data}{a data object}

  \item{covariate.order}{the new order of the covarites}

  \item{...}{additional paramters}
}
\value{
  rearranged sgl data
}
\description{
  todo
}
\author{
  Martin Vincent
}

