\name{rearrange.sgldata}
\alias{rearrange.sgldata}
\title{Rearrange sgldata}
\usage{
  \method{rearrange}{sgldata} (data, covariate.order, ...)
}
\arguments{
  \item{data}{a sgldata object}

  \item{covariate.order}{the new order of the covariates}

  \item{...}{not used}
}
\value{
  a sgldata object with the covariates reordered
}
\description{
  Rearrange the order of the covariates in a sgldata
  object.
}
\author{
  Martin Vincent
}
\seealso{
  Other sgldata: \code{\link{create.sgldata}},
  \code{\link{prepare.args}},
  \code{\link{prepare.args.sgldata}}
}

