\name{rearrange.sgldata}
\alias{rearrange.sgldata}
\title{Rearrange sgldata}
\usage{
  \method{rearrange}{sgldata} (data, covariate.order, ...)
}
\arguments{
  \item{data}{sgldata object}

  \item{covariate.order}{the new order of the covarites}

  \item{...}{not used}
}
\value{
  a sgl data object with the covariates reordered
}
\description{
  Rearrange sgldata
}
\author{
  Martin Vincent
}
\seealso{
  Other sgldata: \code{\link{create.sgldata}},
  \code{\link{prepare.args.sgldata}}
}

