\name{sgl_predict}
\alias{sgl_predict}
\title{Sgl predict}
\usage{
  sgl_predict(module_name, PACKAGE, object, data, ...)
}
\arguments{
  \item{module_name}{}

  \item{PACKAGE}{}

  \item{object}{}

  \item{data}{}

  \item{...}{}
}
\value{
  sgl object
}
\description{
  Sgl predict
}
\author{
  Martin Vincent
}

